evaluate.py
import json
import requests
import time

# Evaluation Logic ---
def run_evaluation():
    with open("tests/questions.json", "r") as f:
        questions = json.load(f)
    
    results = []
    hit_count = 0
    hallucination_count = 0
    refusal_count = 0

    for q in questions:
        response = requests.post("http://127.0.0.1:8000/ask", json={"question": q['q'], "debug": True})
        data = response.json()
        
       
        refused = data['answer'] == "This information is not available in the provided document(s)."
        
        # Metrics Calculation
        if not refused:
            # If answer provided, check if citations exist (Hit Rate)
            if len(data['citations']) > 0: hit_count += 1
            # If an out-of-bounds question (like Q11) was answered, it's a hallucination
            if q['id'] in [11, 20, 23, 28, 40]: hallucination_count += 1
        else:
            refusal_count += 1

        results.append({
            "id": q['id'],
            "type": q['type'],
            "answer": data['answer'],
            "citations": data['citations'],
            "status": "Refused" if refused else "Answered"
        })
        time.sleep(1)

    # Summary Report
    total = len(questions)
    report = {
        "hit_rate": f"{(hit_count/total)*100}%",
        "hallucination_rate": f"{(hallucination_count/total)*100}%",
        "faithfulness": "High (Strict System Prompt used)",
        "results": results
    }

    with open("tests/eval_report.json", "w") as f:
        json.dump(report, f, indent=4)
    print("Level 1 Evaluation Report generated in tests/eval_report.json")

if __name__ == "__main__":
    run_evaluation()