requirements.txt
fastapi
uvicorn
langchain
langchain-text-splitters
langchain-community
langchain-core
langchain-huggingface
langchain-groq
faiss-cpu
pymupdf
python-dotenv
sentence-transformers